Fixed thanks to some work w/ the amazing Rob Winch!
