+++ livelessons-security/livelessons-security-sso-auth-server/src/main/java/demo/SecurityConfig.java
@@ -1,0 +1,37 @@
+package demo;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.authentication.configuration.EnableGlobalAuthentication;
+import org.springframework.security.config.annotation.authentication.configurers.GlobalAuthenticationConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+
+@Configuration
+@EnableGlobalAuthentication
+public class SecurityConfig extends GlobalAuthenticationConfigurerAdapter {
+
+    @Autowired
+    private UserDetailsService userDetailsService;
+
+    @Override
+    public void init(AuthenticationManagerBuilder auth) throws Exception {
+        auth.userDetailsService(this.userDetailsService);
+    }
+
+}
+
+
+/*
+@Configuration
+@EnableGlobalAuthentication
+public class SecurityConfig  {
+
+    @Autowired
+    public void configureGlobal(AuthenticationManagerBuilder auth,
+                                UserDetailsService userDetailsService) throws Exception {
+        auth.userDetailsService(userDetailsService);
+    }
+}
+*/
